fix(#88): shorten relocated type names/paths to satisfy the 100-char limit
