+++ src/modules/user/features/auth/types/auth-forms/notification-shared-styles.ts
