+++ src/modules/user/types/store/login-response-mapper.ts
@@ -3,6 +3,4 @@ import type { LoginResponse } from '@auth/types/api-responses';
 
 export type LoginSuccessPayload = LoginResponse & { email: string };
 
-export type LoginResponseMappingResult =
-  | { ok: true; value: LoginSuccessPayload }
-  | { ok: false; error: UiError };
+export type LoginResult = { ok: true; value: LoginSuccessPayload } | { ok: false; error: UiError };
