+++ src/modules/user/types/store/registration-response-mapper.ts
@@ -1,6 +1,4 @@
 import type { UiError } from '@/services/error';
 import type { SafeUserInfo } from '@auth/types/api-responses';
 
-export type RegistrationResponseMappingResult =
-  | { ok: true; value: SafeUserInfo }
-  | { ok: false; error: UiError };
+export type RegistrationResult = { ok: true; value: SafeUserInfo } | { ok: false; error: UiError };
